feat: Began working on App, forgot to do multiple commits
Forgot to do multiple commits here, but I added an app_old.js to preserve the original js file just for easy reference. I added BookGrid and Bookshelf, each of these is a React Component that im using to generate the shelves and the books, still in the process of coding these all out.